Move plone.outputfilters ImageCaptioning to plone.app.layout
